fix(barrier-method): fix general method and add constraint types

+++ src/vicentin/optimization/minimization/newton_method/newton_torch.py
@@ -119,7 +119,7 @@ def newton(
     return_dual: bool = False,
     return_loss: bool = False,
 ):
-    x = x0.clone().detach().to(torch.float64)
+    x = x0.clone().detach()
     i = 0
     loss = []
